Graphics: Fix Type Issue in Loop

+++ Kernel/Graphics/Graphics.cpp
@@ -11,8 +11,8 @@ void Graphics::init_terminal()
     // Using 0 as a value because we are clearing the terminal
     auto initial_value = ScreenChar { static_cast<std::uint8_t>('a'), ColorCode { VGA_Color::White, VGA_Color::Black } };
 
-    for (std::uint16_t i = 0; i < m_row; ++i) {
-        for (std::uint16_t j = 0; j < m_column; ++j) {
+    for (std::size_t i = 0; i < m_row; ++i) {
+        for (std::size_t j = 0; j < m_column; ++j) {
             *m_buffer[i][j] = initial_value;
             ++initial_value.ascii_char;
             ++m_column_position;
